refactor: rename components
