+++ react-app/.prettierrc.cjs
@@ -1,0 +1,9 @@
+module.exports = {
+    endOfLine: "auto",
+    semi: true,
+    singleQuote: true,
+    tabWidth: 2,
+    trailingComma: "es5",
+    printWidth: 100,
+    arrowParens: "always"
+  };
